@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Browser free DevTools for your command line.",
   "main": "index.js",
+  "type": "module",
   "dependencies": {
     "cli-table3": "^0.6.1",
     "copy-paste-win32fix": "^1.4.0",
@@ -31,10 +32,10 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "eslint": "^7.23.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-import": "^2.22.1",
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.3.1"
+    "eslint-plugin-promise": "^5.2.0"
   }
 }
